Move functionality from attributeSideBar.deltas to metricDeltaSelected #527
