--- visualization/app/codeCharta/ui/attributeSideBar/attributeSideBar.deltas.component.html
@@ -1,3 +1,0 @@
-<span class="metric-delta {{ $ctrl._viewModel.colorClass }}" ng-show="$ctrl._viewModel.deltaValue">
-	&Delta;{{ $ctrl._viewModel.deltaValue | number }}
-</span>
